APPS CON ESTRUCTURA/FORENSIGEN/requirements.txt
huggingface_hub==0.25.2
gradio==4.16.0
groq
python-docx

